Refactor httpSubscription into new module
