+++ src/lib/figma/navigation/side-nav/WorkspaceSelector.svelte
@@ -48,7 +48,11 @@
 
     let workspaceContextMenuOpen = false;
 
-    async function showWorkspaceContextMenu(workspaces: Workspace[]) {
+    async function showWorkspaceContextMenu() {
+        const workspaces = $currentWorkspaces;
+        if (!workspaces) {
+            throw new Error("Expected workspaces");
+        }
         workspaceContextMenuOpen = true;
         try {
             await openContextMenu(
@@ -62,8 +66,6 @@
             workspaceContextMenuOpen = false;
         }
     }
-
-    $: workspaces = $currentWorkspaces;
 </script>
 
 {#if open}
@@ -71,9 +73,8 @@
         <div class="flex flex-row items-center justify-between gap-4">
             <div class="min-w-0 grow" bind:this={workspaceContextMenuAnchor}>
                 <button
-                    disabled={workspaces === undefined}
-                    on:click={workspaces &&
-                        showWorkspaceContextMenu.bind(null, workspaces)}
+                    disabled={$currentWorkspaces === undefined}
+                    on:click={showWorkspaceContextMenu}
                     class="flex w-full flex-row items-center justify-between gap-2 rounded-lg border border-border p-2 hover:bg-secondary-hover"
                 >
                     <div class="flex min-w-0 flex-row items-center gap-2">
@@ -120,8 +121,9 @@
         <div bind:this={workspaceContextMenuAnchor}>
             <BorderedIcon
                 type="workspace"
-                on:click={workspaces &&
-                    showWorkspaceContextMenu.bind(null, workspaces)}
+                on:click={$currentWorkspaces !== undefined
+                    ? showWorkspaceContextMenu
+                    : undefined}
             />
         </div>
         <div bind:this={sideNavContextMenuAnchor}>
